修改初始化深度

+++ AirRenderer/Configuration.cpp
@@ -14,6 +14,6 @@ glm::mat4 Configuration::GetScreenMatrix()
 Configuration::Configuration()
 {
 	this->colorBuffer = new Buffer<Color>(800, 450, 1, Color(0, 0, 0, 1));
-	this->depthBuffer = new Buffer<float>(800, 450, 1, float(1));
+	this->depthBuffer = new Buffer<float>(800, 450, 1, float(999));
     this->resolution = Resolution(800, 450);
 }
